xfs: Set up infastructure for deferred attribute operations
Currently attributes are modified directly across one or more transactions.
But they are not logged or replayed in the event of an error. The goal of delayed
attributes is to enable logging and replaying of attribute operations using the
existing delayed operations infrastructure.  This will later enable the attributes
to become part of larger multi part operations that also must first be recorded to
the log.  This is mostly of interest in the scheme of parent pointers which would
need to maintain an attribute containing parent inode information any time an inode
is moved, created, or removed.  Parent pointers would then be of interest to any
feature that would need to quickly derive an inode path from the mount point.
 Online scrub, nfs lookups and fs grow or shrink operations are all features that
could take advantage of this.

This patch adds two new log item types for setting or removing attributes as
deferred operations.  The xfs_attri_log_item logs an intent to set or remove an
attribute.  The corresponding xfs_attrd_log_item holds a reference to the
xfs_attri_log_item and is freed once the transaction is done.  Both log items use a
generic xfs_attr_log_format structure that contains the attribute name, value,
flags, inode, and an op_flag that indicates if the operations is a set or remove.

Signed-off-by: Allison Collins <allison.henderson@oracle.com>