xfs: define parent pointer xattr format
We need to define the parent pointer attribute format before we start adding support
for it into all the code that needs to use it. The EA format we will use encodes the
following information:

        name={parent inode #, parent inode generation, dirent offset}
        value={dirent filename}

The inode/gen gives all the information we need to reliably identify the parent
without requiring child->parent lock ordering, and allows userspace to do
pathname component level reconstruction without the kernel ever needing to verify
the parent itself as part of ioctl calls.

By using the dirent offset in the EA name, we have a method of knowing the exact
parent pointer EA we need to modify/remove in rename/unlink without an unbound EA
name search.

By keeping the dirent name in the value, we have enough information to be able to
validate and reconstruct damaged directory trees. While the diroffset of a filename
alone is not unique enough to identify the child, the
{diroffset,filename,child_inode} tuple is sufficient. That is, if the diroffset
gets reused and points to a different filename, we can detect that from the contents
of EA. If a link of the same name is created, then we can check whether it points at
the same inode as the parent EA we current have.

[achender: rebased, changed __unint32_t to xfs_dir2_dataptr_t,
           changed p_ino to xfs_ino_t and p_namelen to uint8_t,
           moved to xfs_da_format for xfs_dir2_dataptr_t]

Signed-off-by: Dave Chinner <dchinner@redhat.com>
Signed-off-by: Allison Henderson <allison.henderson@oracle.com>
Reviewed-by: Darrick J. Wong<darrick.wong@oracle.com>