xfs: Add parent pointers to rename
This patch removes the old parent pointer attribute during the
rename operation, and re-adds the updated parent pointer.  In
the case of xfs_cross_rename, we modify the routine not to roll
the transaction just yet.  We will do this after the parent
pointer is added in the calling xfs_rename function.

Signed-off-by: Allison Henderson <allison.henderson@oracle.com>