xfsprogs: implement the upper half of parent pointers
Add ioctl definitions to libxfs, build the necessary helpers into libfrog and
libhandle to iterate parents (and parent paths), then wire up xfs_scrub to be able
to query parent pointers from userspace.  The goal of this patch is to exercise
userspace, and is nowhere near a complete solution.  A basic xfs_io parent command
implementation replaces ... whatever that is that's there now.

[achender: Minor syntax adjustments to sew solution in actual support
       in libxfs for working with parent ptrs]

Signed-off-by: Darrick J. Wong <darrick.wong@oracle.com>
Signed-off-by: Allison Collins <allison.henderson@oracle.com>