RFC xfsprogs: Skip flip flags for delayed attrs
Source kernel commit: 2b1c81a8c3f453ba16b6db8dae256723bf53c051

This is a clean up patch that skips the flip flag logic for delayed attr
renames.  Since the log replay keeps the inode locked, we do not need to
worry about race windows with attr lookups.  So we can skip over
flipping the flag and the extra transaction roll for it

RFC:
In the last review, folks asked for some performance analysis, so I
did a few perf captures with and with out this patch.  What I found was
that there wasnt very much difference at all between having the patch or
not having it.  Of the time we do spend in the affected code, the
percentage is minimal.  Most of the time we spend about %0.03 of the
time in this function, with or with out the patch.  Occasionally we get
a 0.02%, but it seems more intermittent rather than being related to the
change.  So I think this starts to challenge needing this patch at all.
This patch was requested some number of reviews ago, be perhaps in light
of the findings, it may not be as appealing now.

     0.03%     0.00%  fsstress  [xfs]               [k] xfs_attr_set_iter

Shall we drop it, or is there still reason to keep it?

Signed-off-by: Allison Henderson <allison.henderson@oracle.com>
Reviewed-by: Darrick J. Wong <djwong@kernel.org>